pyproject.toml
[project]
name = "AutoTARA-RAG"
version = "0.1.0"
description = "Full multi-agent RAG-enabled automotive TARA pipeline"
authors = [{name="Novatratech LTD"}]
requires-python = ">=3.10"

dependencies = [
    "pydantic==2.6.1",
    "python-dotenv==1.0.1",
    "openai==1.12.0",
    "faiss-cpu==1.7.4",
    "chromadb==0.4.24",
    "streamlit==1.32.0",
    "tqdm==4.66.2",
    "numpy==1.26.4",
    "pandas==2.2.1",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
